feat: 클래스 통합 및 토큰방식 세션으로 변경
UserService AuthService 통합
AuthController UserController 통합
인증방식 토큰 -> 세션으로 변경
