--- src/main/java/com/example/mvctest/config/jwt/JwtProvider.java
@@ -1,55 +1,0 @@
-package com.example.mvctest.config.jwt;
-
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-
-@Slf4j
-@Component
-public class JwtProvider {
-
-    @Value("${jwt.secret:default-secret-key-that-is-long-enough-for-hs256}")
-    private String secretKeyString;
-
-    private Key secretKey;
-
-    private final long tokenValidityInMilliseconds = 1000 * 60 * 60; // 1시간
-
-    @PostConstruct
-    protected void init() {
-        this.secretKey = Keys.hmacShaKeyFor(secretKeyString.getBytes());
-    }
-
-    public String createToken(String email) {
-        Claims claims = Jwts.claims().setSubject(email);
-        Date now = new Date();
-        Date validity = new Date(now.getTime() + tokenValidityInMilliseconds);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(validity)
-                .signWith(secretKey, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public String getEmailFromToken(String token) {
-        return Jwts.parserBuilder().setSigningKey(secretKey).build().parseClaimsJws(token).getBody().getSubject();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder().setSigningKey(secretKey).build().parseClaimsJws(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            log.error("Invalid JWT token: {}", e.getMessage());
-            return false;
-        }
-    }
-}
