+++ src/main/java/com/example/mvctest/controller/UserController.java
@@ -1,7 +1,9 @@
 package com.example.mvctest.controller;
 
 import com.example.mvctest.dto.UserLoginRequestDTO;
+import com.example.mvctest.dto.UserResponseDTO;
 import com.example.mvctest.dto.UserSignupRequestDTO;
+import com.example.mvctest.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -10,21 +12,22 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Collections;
-import java.util.Map;
-
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
 public class UserController {
 
+    private final UserService userService;
+
     @PostMapping("/signup")
-    public ResponseEntity<Map<String, String>> signup(@Valid @RequestBody UserSignupRequestDTO userSignupRequestDTO){
-        return ResponseEntity.ok(Collections.singletonMap("message", "회원가입 성공"));
+    public ResponseEntity<String> signUp(@Valid @RequestBody UserSignupRequestDTO requestDto) {
+        userService.signUp(requestDto);
+        return ResponseEntity.ok("User signed up successfully");
     }
 
     @PostMapping("/login")
-    public ResponseEntity<Map<String, String>> login(@Valid @RequestBody UserLoginRequestDTO userLoginRequestDTO){
-        return ResponseEntity.ok(Collections.singletonMap("message", "로그인 성공"));
+    public ResponseEntity<UserResponseDTO> login(@Valid @RequestBody UserLoginRequestDTO requestDto) {
+        UserResponseDTO responseDto = userService.login(requestDto);
+        return ResponseEntity.ok(responseDto);
     }
 }
